move dirtyEnhance, I gave up typing
