+++ frontend/src/lib/RelationEditor.svelte
@@ -3,7 +3,7 @@
 	import client from './api';
 	import { SongRelationPredicate, WorkRelationEditorPredicate } from './enums';
 	import { m } from './paraglide/messages';
-	import { dirtyEnhance } from './ui';
+	import { dirtyEnhance } from '$lib/dirty';
 	import { getLocale } from './paraglide/runtime';
 	import type { components } from './schema';
 	import SongField from './SongField.svelte';
